cluster_finder/requirements.txt
numpy>=1.21
matplotlib>=3.4
pymatgen>=2023.0
networkx>=2.6
pandas>=1.3
mp-api>=0.33.3
pytest>=7.4.0,<9.0.0
pydantic>=1.10.8,<3.0.0
ase==3.22.1
aiohttp>=3.8.0,<4.0.0
asyncio>=3.4.3,<4.0.0
typer>=0.9.0
rich>=13.0.0
joblib>=1.3.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
scipy>=1.7
tqdm>=4.65.0